Breaking: Rename cdp connector to chrome
Work on #361
