+++ tests/lib/config.ts
@@ -44,20 +44,20 @@ test(`if package.json is an invalid JSON, it should return an exception`, (t) =>
 test(`if the config file doesn't have an extension, it should be parse as JSON file`, (t) => {
     const configuration = config.load(path.join(__dirname, './fixtures/sonarrc'));
 
-    t.is((configuration.connector as IConnectorConfig).name, 'cdp');
+    t.is((configuration.connector as IConnectorConfig).name, 'chrome');
     t.is(configuration.rules['disallowed-headers'], 'warning');
 });
 
 test(`if the config file is JavaScript, it should return the configuration part`, (t) => {
     const configuration = config.load(path.join(__dirname, './fixtures/sonarrc.js'));
 
-    t.is((configuration.connector as IConnectorConfig).name, 'cdp');
+    t.is((configuration.connector as IConnectorConfig).name, 'chrome');
     t.is(configuration.rules['disallowed-headers'], 'warning');
 });
 
 test(`if package.json contains a valid sonar coniguration, it should return it`, (t) => {
     const configuration = config.load(path.join(__dirname, './fixtures/package.json'));
 
-    t.is((configuration.connector as IConnectorConfig).name, 'cdp');
+    t.is((configuration.connector as IConnectorConfig).name, 'chrome');
     t.is(configuration.rules['disallowed-headers'], 'warning');
 });
